rpl: merge storing/non-storing to reduce the amount of redundant code
