+++ sys/net/routing/rpl/rpl.c
@@ -1,5 +1,6 @@
 /*
- * Copyright (C) 2013, 2014  INRIA.
+ * Copyright (C) 2013 - 2014  INRIA.
+ * Copyright (C) 2015 Cenk Gündoğan <cnkgndgn@gmail.com>
  *
  * This file is subject to the terms and conditions of the GNU Lesser General
  * Public License v2.1. See the file LICENSE in the top level directory for more
@@ -18,6 +19,7 @@
  *
  * @author      Eric Engel <eric.engel@fu-berlin.de>
  * @author      Fabian Brandt <fabianbr@zedat.fu-berlin.de>
+ * @author      Cenk Gündoğan <cnkgndgn@gmail.com>
  */
 
 #include <string.h>
@@ -33,14 +35,6 @@
 
 #include "sixlowpan.h"
 #include "net_help.h"
-
-#if RPL_DEFAULT_MOP == RPL_MOP_STORING_MODE_NO_MC
-#include "rpl/rpl_storing.h"
-#elif RPL_DEFAULT_MOP == RPL_MOP_NON_STORING_MODE
-#include "rpl/rpl_nonstoring.h"
-#else
-#include "rpl/rpl_storing.h"
-#endif
 
 #define ENABLE_DEBUG (0)
 #if ENABLE_DEBUG
@@ -73,7 +67,7 @@ static msg_t srh_m_send, srh_m_recv;
 static rpl_routing_entry_t rpl_routing_table[RPL_MAX_ROUTING_ENTRIES];
 #endif
 uint8_t rpl_max_routing_entries;
-static ipv6_addr_t my_address;
+ipv6_addr_t my_address;
 
 /* IPv6 message buffer */
 static ipv6_hdr_t *ipv6_buf;
@@ -106,33 +100,11 @@ uint8_t rpl_init(int if_id)
 
     /* initialize objective function manager */
     rpl_of_manager_init(&my_address);
-    rpl_init_mode(&my_address);
 
     rt_time = timex_set(RPL_LIFETIME_STEP, 0);
     vtimer_set_msg(&rt_timer, rt_time, rpl_process_pid, RPL_MSG_TYPE_ROUTING_ENTRY_UPDATE, NULL);
 
     return SIXLOWERROR_SUCCESS;
-}
-
-void rpl_init_root(void)
-{
-#if (RPL_DEFAULT_MOP == RPL_MOP_NON_STORING_MODE)
-#ifndef RPL_NODE_IS_ROOT
-puts("\n############################## ERROR ###############################");
-puts("This configuration has NO ROUTING TABLE available for the root node!");
-puts("The root will NOT be INITIALIZED.");
-puts("Please build the binary for root in non-storing MOP with:");
-puts("\t\t'make RPL_NODE_IS_ROOT=1'");
-puts("############################## ERROR ###############################\n");
-return;
-#endif
-#endif
-    rpl_init_root_mode();
-}
-
-uint8_t rpl_is_root(void)
-{
-    return rpl_is_root_mode();
 }
 
 #if (RPL_DEFAULT_MOP == RPL_MOP_NON_STORING_MODE)
@@ -154,7 +126,7 @@ void internal_srh_process(ipv6_srh_t *srh_header)
         srh_header->segments_left = segs - 1;
         DEBUGF("Segments left after reduction: %d\n", srh_header->segments_left);
         DEBUGF("Next hop is: %s\n",
-                ipv6_addr_to_str(addr_str, IPV6_MAX_ADDR_STR_LEN, down_next_hop));
+               ipv6_addr_to_str(addr_str, IPV6_MAX_ADDR_STR_LEN, down_next_hop));
     }
 }
 #endif
@@ -176,9 +148,11 @@ void *rpl_process(void *arg)
             switch (m_recv.type) {
                 case RPL_MSG_TYPE_DAO_HANDLE:
                     dodag = (rpl_dodag_t *) m_recv.content.ptr;
+
                     if (dodag->joined) {
                         _dao_handle_send(dodag);
                     }
+
                     break;
 
                 case RPL_MSG_TYPE_ROUTING_ENTRY_UPDATE:
@@ -187,16 +161,20 @@ void *rpl_process(void *arg)
 
                 case RPL_MSG_TYPE_TRICKLE_INTERVAL:
                     trickle = (trickle_t *) m_recv.content.ptr;
+
                     if (trickle->callback.func != NULL) {
                         trickle_interval(trickle);
                     }
+
                     break;
 
                 case RPL_MSG_TYPE_TRICKLE_CALLBACK:
                     trickle = (trickle_t *) m_recv.content.ptr;
+
                     if (trickle->callback.func != NULL) {
                         trickle_callback(trickle);
                     }
+
                     break;
 
                 default:
@@ -214,7 +192,7 @@ void *rpl_process(void *arg)
 
                 /* get code for message-interpretation and process message */
                 DEBUGF("Received RPL information of type %04X and length %u\n",
-                        m_recv.type, NTOHS(ipv6_buf->length));
+                       m_recv.type, NTOHS(ipv6_buf->length));
 
                 switch (m_recv.type) {
                     case (ICMP_CODE_DIS): {
@@ -252,21 +230,23 @@ void *rpl_process(void *arg)
                     /* if there are no segments left, the routing is finished */
                     if (srh_header->segments_left == 0) {
                         DEBUGF("Source routing finished with next header: %02X.\n",
-                                srh_header->nextheader);
+                               srh_header->nextheader);
                         DEBUGF("Size of srh: %d\n", srh_header->hdrextlen);
                         uint8_t *payload = ((uint8_t *)(m_recv.content.ptr +
-                                    IPV6_HDR_LEN + sizeof(ipv6_srh_t)+srh_header->hdrextlen));
+                                                        IPV6_HDR_LEN + sizeof(ipv6_srh_t) + srh_header->hdrextlen));
                         rpl_remove_srh_header(ipv6_buf, payload, srh_header->nextheader);
                     }
                     else {
                         internal_srh_process(srh_header);
+
                         if (down_next_hop != NULL) {
                             uint8_t *payload = ((uint8_t *)(m_recv.content.ptr + IPV6_HDR_LEN));
                             rpl_srh_sendto(payload, NTOHS(ipv6_buf->length), &ipv6_buf->srcaddr,
-                                    down_next_hop, srh_header, 0);
+                                           down_next_hop, srh_header, 0);
                         }
                     }
                 }
+
 #if RPL_MAX_ROUTING_ENTRIES != 0
                 else  {
                     srh_header = rpl_get_srh_header(ipv6_buf);
@@ -274,87 +254,24 @@ void *rpl_process(void *arg)
                     if (srh_header != NULL) {
                         uint8_t *payload = ((uint8_t *)(m_recv.content.ptr + IPV6_HDR_LEN));
                         rpl_srh_sendto(payload, NTOHS(ipv6_buf->length),
-                                &ipv6_buf->srcaddr, &ipv6_buf->destaddr, srh_header,
-                                srh_header->hdrextlen + sizeof(ipv6_srh_t));
+                                       &ipv6_buf->srcaddr, &ipv6_buf->destaddr, srh_header,
+                                       srh_header->hdrextlen + sizeof(ipv6_srh_t));
                     }
                 }
+
 #endif
             }
+
 #endif
         }
     }
 }
 
-void rpl_send_DIO(ipv6_addr_t *destination)
+void _rpl_update_routing_table(void)
 {
-    if (destination) {
-        DEBUGF("Send DIO to %s\n", ipv6_addr_to_str(addr_str, IPV6_MAX_ADDR_STR_LEN, destination));
-    }
-
-    rpl_send_DIO_mode(destination);
-}
-
-void rpl_send_DAO(ipv6_addr_t *destination, uint8_t lifetime,
-        bool default_lifetime, uint8_t start_index)
-{
-    if (destination) {
-        DEBUGF("Send DAO to %s\n", ipv6_addr_to_str(addr_str, IPV6_MAX_ADDR_STR_LEN, destination));
-    }
-
-    rpl_send_DAO_mode(destination, lifetime, default_lifetime, start_index);
-}
-
-void rpl_send_DIS(ipv6_addr_t *destination)
-{
-    if (destination) {
-        DEBUGF("Send DIS to %s\n", ipv6_addr_to_str(addr_str, IPV6_MAX_ADDR_STR_LEN, destination));
-    }
-
-    rpl_send_DIS_mode(destination);
-}
-
-void rpl_send_DAO_ACK(ipv6_addr_t *destination)
-{
-    if (destination) {
-        DEBUGF("Send DAO ACK to %s\n",
-                ipv6_addr_to_str(addr_str, IPV6_MAX_ADDR_STR_LEN, destination));
-    }
-
-    rpl_send_DAO_ACK_mode(destination);
-}
-
-void rpl_recv_DIO(void)
-{
-    DEBUGF("DIO received\n");
-
-    rpl_recv_DIO_mode();
-}
-
-void rpl_recv_DAO(void)
-{
-    DEBUGF("DAO received\n");
-
-    rpl_recv_DAO_mode();
-}
-
-void rpl_recv_DIS(void)
-{
-    DEBUGF("DIS received\n");
-
-    rpl_recv_DIS_mode();
-
-}
-
-void rpl_recv_DAO_ACK(void)
-{
-    DEBUGF("DAO ACK received\n");
-
-    rpl_recv_dao_ack_mode();
-}
-
-void _rpl_update_routing_table(void) {
     rpl_dodag_t *my_dodag = rpl_get_my_dodag();
     rpl_routing_entry_t *rt;
+
     if (my_dodag != NULL) {
         rt = rpl_get_routing_table();
 
@@ -393,7 +310,7 @@ void rpl_delay_dao(rpl_dodag_t *dodag)
     dodag->ack_received = false;
     vtimer_remove(&dodag->dao_timer);
     vtimer_set_msg(&dodag->dao_timer, dodag->dao_time,
-            rpl_process_pid, RPL_MSG_TYPE_DAO_HANDLE, dodag);
+                   rpl_process_pid, RPL_MSG_TYPE_DAO_HANDLE, dodag);
 }
 
 /* This function is used for regular update of the routes.
@@ -406,7 +323,7 @@ void long_delay_dao(rpl_dodag_t *dodag)
     dodag->ack_received = false;
     vtimer_remove(&dodag->dao_timer);
     vtimer_set_msg(&dodag->dao_timer, dodag->dao_time,
-            rpl_process_pid, RPL_MSG_TYPE_DAO_HANDLE, dodag);
+                   rpl_process_pid, RPL_MSG_TYPE_DAO_HANDLE, dodag);
 }
 
 void rpl_dao_ack_received(rpl_dodag_t *dodag)
@@ -415,14 +332,15 @@ void rpl_dao_ack_received(rpl_dodag_t *dodag)
     long_delay_dao(dodag);
 }
 
-void _dao_handle_send(rpl_dodag_t *dodag) {
+void _dao_handle_send(rpl_dodag_t *dodag)
+{
     if ((dodag->ack_received == false) && (dodag->dao_counter < DAO_SEND_RETRIES)) {
         dodag->dao_counter++;
         rpl_send_DAO(NULL, 0, true, 0);
         dodag->dao_time = timex_set(DEFAULT_WAIT_FOR_DAO_ACK, 0);
         vtimer_remove(&dodag->dao_timer);
         vtimer_set_msg(&dodag->dao_timer, dodag->dao_time,
-                rpl_process_pid, RPL_MSG_TYPE_DAO_HANDLE, dodag);
+                       rpl_process_pid, RPL_MSG_TYPE_DAO_HANDLE, dodag);
     }
     else if (dodag->ack_received == false) {
         long_delay_dao(dodag);
@@ -433,32 +351,34 @@ ipv6_addr_t *rpl_get_next_hop(ipv6_addr_t *addr)
 {
 
     DEBUGF("Looking up the next hop to %s\n",
-            ipv6_addr_to_str(addr_str, IPV6_MAX_ADDR_STR_LEN, addr));
+           ipv6_addr_to_str(addr_str, IPV6_MAX_ADDR_STR_LEN, addr));
 
 #if RPL_MAX_ROUTING_ENTRIES != 0
+
     for (uint8_t i = 0; i < rpl_max_routing_entries; i++) {
         if (rpl_routing_table[i].used) {
             DEBUGF("checking %d: %s\n", i,
-                    ipv6_addr_to_str(addr_str, IPV6_MAX_ADDR_STR_LEN, &rpl_routing_table[i].address));
+                   ipv6_addr_to_str(addr_str, IPV6_MAX_ADDR_STR_LEN, &rpl_routing_table[i].address));
         }
 
         if ((RPL_DEFAULT_MOP == RPL_MOP_NON_STORING_MODE) && rpl_is_root()) {
             if (rpl_routing_table[i].used && rpl_equal_id(&rpl_routing_table[i].address, addr)) {
                 DEBUGF("found %d: %s\n", i,
-                        ipv6_addr_to_str(addr_str, IPV6_MAX_ADDR_STR_LEN,
-                            &rpl_routing_table[i].address));
+                       ipv6_addr_to_str(addr_str, IPV6_MAX_ADDR_STR_LEN,
+                                        &rpl_routing_table[i].address));
                 return &rpl_routing_table[i].address;
             }
         }
         else {
             if (rpl_routing_table[i].used && rpl_equal_id(&rpl_routing_table[i].address, addr)) {
                 DEBUGF("found %d: %s\n", i,
-                        ipv6_addr_to_str(addr_str, IPV6_MAX_ADDR_STR_LEN,
-                            &rpl_routing_table[i].next_hop));
+                       ipv6_addr_to_str(addr_str, IPV6_MAX_ADDR_STR_LEN,
+                                        &rpl_routing_table[i].next_hop));
                 return &rpl_routing_table[i].next_hop;
             }
         }
     }
+
 #else
     (void) addr;
 #endif
@@ -570,9 +490,9 @@ void rpl_add_srh_entry(ipv6_addr_t *child, ipv6_addr_t *parent, uint16_t lifetim
      * so that route aggregation can be done properly.
      */
     DEBUGF("Adding source-routing entry child: %s\n",
-            ipv6_addr_to_str(addr_str, IPV6_MAX_ADDR_STR_LEN, child));
+           ipv6_addr_to_str(addr_str, IPV6_MAX_ADDR_STR_LEN, child));
     DEBUGF("Adding source-routing entry parent: %s\n",
-            ipv6_addr_to_str(addr_str, IPV6_MAX_ADDR_STR_LEN, parent));
+           ipv6_addr_to_str(addr_str, IPV6_MAX_ADDR_STR_LEN, parent));
 
     for (uint8_t i = 0; i < rpl_max_routing_entries; i++) {
         if (!rpl_routing_table[i].used) {
@@ -621,12 +541,12 @@ ipv6_srh_t *rpl_get_srh_header(ipv6_hdr_t *act_ipv6_hdr)
 
         for (uint8_t i = 0; i < rpl_max_routing_entries; i++) {
             if (rpl_routing_table[i].used
-                    && ipv6_suffix_is_equal(&rpl_routing_table[i].address, actual_node)) {
+                && ipv6_suffix_is_equal(&rpl_routing_table[i].address, actual_node)) {
                 DEBUGF("[INFO] Found parent-child relation with P: %s\n",
-                        ipv6_addr_to_str(addr_str, IPV6_MAX_ADDR_STR_LEN,
-                            &rpl_routing_table[i].next_hop));
+                       ipv6_addr_to_str(addr_str, IPV6_MAX_ADDR_STR_LEN,
+                                        &rpl_routing_table[i].next_hop));
                 DEBUGF(" and C: %s\n",
-                        ipv6_addr_to_str(addr_str, IPV6_MAX_ADDR_STR_LEN, actual_node));
+                       ipv6_addr_to_str(addr_str, IPV6_MAX_ADDR_STR_LEN, actual_node));
                 memcpy(&rev_route[counter], actual_node, sizeof(ipv6_addr_t));
 
                 actual_node = &rpl_routing_table[i].next_hop;
@@ -638,6 +558,7 @@ ipv6_srh_t *rpl_get_srh_header(ipv6_hdr_t *act_ipv6_hdr)
                     DEBUGF("Error with computing source routing header.\n");
                     return NULL;
                 }
+
                 break;
             }
         }
@@ -652,11 +573,12 @@ ipv6_srh_t *rpl_get_srh_header(ipv6_hdr_t *act_ipv6_hdr)
      * After building it starts with the node next to destination
      */
     if (counter > 1) {
-        for (uint8_t i = 0; i < counter-1; i++) {
-            memcpy(&srh_header->route[i], &rev_route[counter-i-2], sizeof(ipv6_addr_t));
+        for (uint8_t i = 0; i < counter - 1; i++) {
+            memcpy(&srh_header->route[i], &rev_route[counter - i - 2], sizeof(ipv6_addr_t));
         }
-        srh_header->hdrextlen = sizeof(ipv6_addr_t)*(counter-1);
-        memcpy(&(act_ipv6_hdr->destaddr), &rev_route[counter-1], sizeof(ipv6_addr_t));
+
+        srh_header->hdrextlen = sizeof(ipv6_addr_t) * (counter - 1);
+        memcpy(&(act_ipv6_hdr->destaddr), &rev_route[counter - 1], sizeof(ipv6_addr_t));
         DEBUGF("Route size: %d\n", srh_header->hdrextlen);
     }
     else {
@@ -689,13 +611,13 @@ void rpl_remove_srh_header(ipv6_hdr_t *ipv6_header, const void *buf, uint8_t nex
     memcpy(payload, buf, msg_length);
     DEBUGF("Source routing header extraction finished.\n");
     DEBUGF("Dest is now: %s\n",
-            ipv6_addr_to_str(addr_str, IPV6_MAX_ADDR_STR_LEN, &temp_ipv6_header->destaddr));
+           ipv6_addr_to_str(addr_str, IPV6_MAX_ADDR_STR_LEN, &temp_ipv6_header->destaddr));
     srh_m_send.content.ptr = (char *) srh_send_buffer;
     msg_send_receive(&srh_m_send, &srh_m_recv, ip_process_pid);
 }
 
 int rpl_srh_sendto(const void *buf, uint16_t len,
-        ipv6_addr_t *src, ipv6_addr_t *dest, ipv6_srh_t *srh_header, uint8_t srh_length)
+                   ipv6_addr_t *src, ipv6_addr_t *dest, ipv6_srh_t *srh_header, uint8_t srh_length)
 {
     ipv6_hdr_t *temp_ipv6_header = ((ipv6_hdr_t *)(&srh_send_buffer));
     ipv6_srh_t *current_packet = ((ipv6_srh_t *)(&srh_send_buffer[IPV6_HDR_LEN]));
@@ -711,6 +633,6 @@ int rpl_srh_sendto(const void *buf, uint16_t len,
     DEBUGF("My payload length: %d\n", plength);
 
     return ipv6_sendto(&temp_ipv6_header->destaddr, IPV6_PROTO_NUM_SRH,
-            (uint8_t *)current_packet, plength, &temp_ipv6_header->destaddr);
+                       (uint8_t *)current_packet, plength, &temp_ipv6_header->destaddr);
 }
 #endif
