--- sys/net/routing/rpl/rpl_nonstoring/Makefile
@@ -1,3 +1,0 @@
-MODULE := rpl
-INCLUDES += -I$(RIOTBASE)/sys/net/routing/rpl
-include $(RIOTBASE)/Makefile.base
